chore: Move spec files package core (second part) (#772)
* move it around

* rename

* make git guardian happy

* linter
