+++ packages/core/src/connectivity/scp-cf/csrf-token-header.spec.ts
@@ -1,19 +1,14 @@
 import { createLogger } from '@sap-cloud-sdk/util';
-import nock = require('nock');
-import { buildCsrfHeaders } from '../../../src/connectivity/scp-cf';
-import {
-  createCreateRequest,
-  createGetAllRequest,
-  createUpdateRequest
-} from '../../test-util/create-requests';
+import nock from 'nock';
 import {
   defaultBasicCredentials,
   defaultDestination,
   defaultHost,
   mockHeaderRequest
-} from '../../test-util/request-mocker';
-import { addCsrfTokenAndCookies } from '../../../src/connectivity/scp-cf/legacy-csrf-token-header';
-import { Destination } from '../../../src';
+} from '../../../test/test-util/request-mocker';
+import { createCreateRequest } from '../../../test/test-util/create-requests';
+import { buildCsrfHeaders } from './csrf-token-header';
+import { Destination } from './destination';
 
 const standardHeaders = {
   accept: 'application/json',
@@ -21,7 +16,7 @@ const standardHeaders = {
   'content-type': 'application/json'
 };
 
-describe('csrf-token-header', () => {
+describe('buildCsrfHeaders', () => {
   const logger = createLogger('csrf-token-header');
 
   it('should build "cookie" and "x-csrf-token" properties.', async () => {
@@ -83,71 +78,6 @@ describe('csrf-token-header', () => {
     expect(warnSpy).toBeCalledWith(
       'CSRF header response does not include cookies.'
     );
-  });
-
-  describe('[deprecated]', () => {
-    it('should return back header, when headers include csrf token', async () => {
-      const createRequest = createUpdateRequest(defaultDestination);
-      const headerWithCsrf = { params: 'any', 'x-csrf-token': 'defined' };
-      const actual = await addCsrfTokenAndCookies(
-        createRequest,
-        headerWithCsrf
-      );
-      expect(actual).toEqual(headerWithCsrf);
-    });
-
-    it('should return back header, on GET request', async () => {
-      const createRequest = createGetAllRequest(defaultDestination);
-      const getHeaders = { params: 'any' };
-      const actual = await addCsrfTokenAndCookies(createRequest, getHeaders);
-      expect(actual).toEqual(getHeaders);
-    });
-
-    it('should call getCsrfToken() to define "cookie" and "x-csrf-token" properties.', async () => {
-      const request = createCreateRequest(defaultDestination);
-      const mockedHeaders = {
-        'x-csrf-token': 'mocked-x-csrf-token',
-        'set-cookie': ['mocked-cookie-0;mocked-cookie-1', 'mocked-cookie-2']
-      };
-
-      mockHeaderRequest({ request });
-
-      const expected = {
-        cookie: 'mocked-cookie-0;mocked-cookie-2',
-        'x-csrf-token': mockedHeaders['x-csrf-token'],
-        ...standardHeaders
-      };
-      const headers = await addCsrfTokenAndCookies(request, standardHeaders);
-      expect(headers).toEqual(expected);
-    });
-
-    it('"x-csrf-token" should not be defined in header when not defined in CSRF headers response.', async () => {
-      const request = createCreateRequest(defaultDestination);
-
-      mockHeaderRequest({
-        request,
-        responseHeaders: {
-          'set-cookie': ['mocked-cookie-0;mocked-cookie-1', 'mocked-cookie-2']
-        }
-      });
-
-      const actual = await addCsrfTokenAndCookies(request, standardHeaders);
-
-      expect('x-csrf-token' in actual).toBeFalsy();
-    });
-
-    it('"cookie" should not be defined in header when not defined in CSRF headers response.', async () => {
-      const request = createCreateRequest(defaultDestination);
-
-      mockHeaderRequest({
-        request,
-        responseHeaders: { 'x-csrf-token': 'mocked-x-csrf-token' }
-      });
-
-      const actual = await addCsrfTokenAndCookies(request, standardHeaders);
-
-      expect('cookie' in actual).toBeFalsy();
-    });
   });
 
   it('should redirect csrf request when using proxy', async () => {
