+++ packages/core/src/odata-v4/uri-conversion/odata-uri.spec.ts
@@ -7,17 +7,15 @@ import {
   testFilterString,
   testFilterStringV4,
   testNestedFilterLambdaExpressionOnLink
-} from '../../../test-util/filter-factory';
-import { TestEntity } from '../../../test-util/test-services/v4/test-service';
-import {
-  oDataUriV4,
-  filterFunctionsV4,
-  filterFunctionV4
-} from '../../../../src';
+} from '../../../test/test-util/filter-factory';
+import { TestEntity } from '../../../test/test-util/test-services/v4/test-service';
+import { filterFunctionsV4 } from '../filter-functions';
+import { filterFunctionV4 } from '../filter-function';
+import { oDataUriV4 } from './odata-uri';
 
 const { getFilter } = oDataUriV4;
 
-describe('get filters', () => {
+describe('getFilter', () => {
   it('for simple filters', () => {
     expect(getFilter(testFilterStringV4.filter, TestEntity).filter).toBe(
       `${testFilterString.odataStr}`
